Initial commit
Assignment 3: Static Visualization in D3 (credit/no credit)

In this assignment you will use JavaScript and D3 to create a static visualization that explores data that you’ve collected from your life.  Possible examples include (but are not limited to) measuring your heart rate before and after doing jumping jacks, the results of flipping a coin (or several different coins), data collected from an activity tracker such as a Fitbit, or the number of cups of coffee you drink throughout the week. The activity should produce approximately 20 - 30 data points that you can use for your static visualization. The primary goal of the assignment is the creation of the visualization using D3, and the data collection process should ideally feel fun and not strenuous. That said, feel free to be creative in the type of data collected. Just remember, this is an ungraded assignment.

Post your finished piece on the I School web site or elsewhere on the web (e.g., GitHub, JSFiddle), and turn in a screen capture with the URL. (Instructions for using your I School web site are under Coursework/Additional Resources.)  Make sure the live version is accessible to your instructor. No additional text or report is necessary for this assignment.